Changed and modified the command to test
